Change to retention tab cutouts of switch holders
Changed the shape of the cutouts in the switch holders for the retention tabs on switches from being wide open at the bottom to closed with a 60 degree chamfer. The shape resembles a divot, and should be an overall improvement to the printability of the generated keyboards, reducing the number of islands and overhangs requiring support.